Add transformation for primitive port types
The following mapping is used for the transformation:
s::Integer -> a::SignedDiscreteType
s::Real -> a::ContinuousType
s::Boolean -> a::BooleanType

s::String and s::UnlimitedNatural are ignored at the moment.

ASEM primitive types are stored in an a::PrimitiveTypeRepository which
will be initialized if a s::PrimitiveType was added to the SysML model.