Add part transformation (reactions) which causes the bug
The message type of the ASEM message which represents the SysML part
reference will be created correctly. But loading the ASEM resource in
the test cases will load a AESM resource with a wrong reference of the
message type.

Note: The ASEM message element is the wrong element for this kind of 
part reference mapping, but it is used at the moment because no other
ASEM element exists in the meta model which could be added to a ASEM
components typed elements attribute.