Add rename transformation for named elements
The named element transformation will handle all name changes of a named
element which do not need further modifications. Therefore the port name
transformation is the only name change which will be handled at the
moment. The name change of a SysML block requires some persistence
modifications, too. So this kind of name transformation will be handled
separately.